3d delta movement -> {   x = -sin(y) * cos(x)   y = -sin(x)   z = cos(y) * cos(x) }

+++ src/main/java/net/tabby/florafaunarebalance/util/Math/Mh.java
@@ -10,7 +10,7 @@ public class Mh { //# damnit intelIj, why not allow unicode chars....
     public static float rad(float degree) {
         return degree * piOver180;
     }
-    public static double frac(double y) {
+    public static float frac(float y) {
         return y - Mth.floor(y);
     }
